stylegan-stack/requirements.txt
# StyleGAN2-ADA Dependencies for Mono Latent Walk Generator
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
click>=8.0.0
scipy>=1.7.0
psutil>=5.8.0
ninja>=1.10.0

# Additional dependencies that may be needed for StyleGAN2-ADA
requests>=2.25.0
tqdm>=4.60.0